Jan_DemoClientAdded
Added a simple demo client. Basically just copied the dummy-client from the Dialogs-Library in the ServerLib.
Changed the Response-type of the the GetSRs-WebserviceMethod to pure XML to easily convert the body into Data.
Noticed that the Client.lvlib does not show up in the VI Hierachy window. Will research this.