Jan_WaitDialogForOracleAdded
Added a Wait Dialog for Application Shutdown. Is Oracle Loop is still in Progress a progressbar will show for 60 seconds informing the user that Oracle needs to finish its query. If it takes longer than 60 seconds a Error will appear, telling the user to abort the application via task manager.